Unify chfl_trajectory_path with other functions getting string values
All the other functions make a copy of the string (chfl_atom_name, chfl_selection_string),
this one did it differently for no good reason (it is not a performance sensitive function)

+++ tests/doc/capi/chfl_trajectory/memory_buffer.c
@@ -15,7 +15,7 @@ int main() {
     }
 
     const char* result = NULL;
-    size_t size_of_result;
+    uint64_t size_of_result;
     chfl_trajectory_memory_buffer(trajectory, &result, &size_of_result);
 
     chfl_trajectory_close(trajectory);
